requirements.txt
pandas==2.2.3
pyTelegramBotAPI==4.26.0
python-dotenv==1.0.1